requirements.txt
requests
m3u8
ffmpeg-python
tweepy
wget